add parameters for -UseBasicParsing and ruby version to install
one particular use case gave me the idea for this:
I start up a Windows VM and I want to jump right in, install ruby, and
get developing. So, I run this script and find out that I have to start
Internet Explorer and accept the security settings. Then I run the
script again and now I have to walk through an interactive cli, which is
great but what would be really great would be to eventually run this
script all by itself before I even log in (or something like that).

This change adds a commandline parameter for using -UseBasicParsing with
Invoke-WebRequest, which circumvents the need for IE to have been started, and a parameter
to designate a single Ruby version and run a non-interactive install of
the single version.